Removes tests from local bgs, and adds tests to each service; records vcr cassettes.
